add logger + delete method update how i never to use + delete  extend Serializable + add ABSTRACT DAO ONLY JDBC USING

+++ src/main/java/mate/academy/model/Train.java
@@ -34,7 +34,7 @@ public class Train {
     private String name;
 
     @Column
-    private Integer number;
+    private long number;
 
     @Column
     private String adressFrom;
